requirements.txt
annotated-types==0.6.0
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.7.2
beautifulsoup4==4.12.2
blinker==1.6.2
Brotli==1.1.0
bs4==0.0.1
certifi==2023.7.22
cffi==1.16.0
chardet==3.0.4
charset-normalizer==3.3.0
click==8.1.7
distro==1.9.0
feedparser==6.0.10
Flask==2.3.3
Flask-Compress==1.14
Flask-gzip==0.2
Flask-HTTPAuth==4.8.0
Flask-Login==0.6.2
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
googletrans==2.4.0
greenlet==3.0.3
h11==0.14.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.1.5
httpcore==1.0.2
httpx==0.26.0
hyperframe==5.2.0
idna==2.10
itsdangerous==2.1.2
Jinja2==3.1.2
lxml==5.1.0
MarkupSafe==2.1.3
openai==1.10.0
passlib==1.7.4
pillow==10.2.0
pycparser==2.21
pydantic==2.6.0
pydantic_core==2.16.1
PyMySQL==1.1.0
pytz==2023.3.post1
requests==2.31.0
rfc3986==1.5.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.25
tqdm==4.66.1
typing_extensions==4.9.0
tzlocal==5.1
Unidecode==1.3.7
urllib3==2.0.6
Werkzeug==2.3.7
WTForms==3.1.2
